fix(decorators): reconnected components create new subscriptions
The old implementation did not work if a component was removed from the DOM and then later added
again. `componentWillLoad` is only invoked once, but `componentDidUnload` is invoked whenever the
component is removed from the DOM. This created an issue where components tried to unsubscribe when
there were no subscriptions created. The correct way to handle this is to use `connectedCallback`
and `disconnectedCallback` instead and make sure that components subscribe to state changes again
when they are added to the DOM again.

fix Lundalogik/crm-feature#1292