feat(interface/limeobject): improve working with limeobjects
At the moment it's a bit hard to work with the limeobjects that get returned from the @Limeobjects
decorator. The data is returned in its raw form, so if you want to get any relations of the object
you have to manually look them up. Because of this you always have to subscribe to the entire state,
so even if you only want state updates for one object or objects of a specific limetype you will get
updates if any object is updated. Methods like getRelation and getLimetype added to the Limeobject
interface

fix #239