feat(limeobject): deprecate getRelation in favor of getValue
`getRelation` was a bit limited since it could only get the value of a relation and it was also not
possible to get the value of nested properties. The new `getValue` method should fix this. It will
allow any value on the object to be returned and will also allow nested values on related objects to
be returned.

re Lundalogik/crm-feature#1621